Interfaces for specific HTTP methods with mutable or immutable access.
The main point of this change is to add methods (get, post, put, and
delete) that are slightly more typesafe than `at`, for registering
handlers for routes with specific HTTP methods. Instead of passing
the top-level state to the handlers for these routes, we treat the
state as something that acts like a RwLock, automatically acquire
the lock for reading or writing, depending on the method and then
pass either a &State or &mut State to the handler. This decreases
boilerplate and makes it harder to accidentally register a mutating
route handler for a pure route, like GET.

In addition, this is a pretty big refactor of the original interfaces
in order to clean up and settle on a pattern for dealing with async
functions and lifetimes.

* By using associated types to name or uniquely identify some types
  in function/trait bounds, I was able to eliminate a few phantom
  type parameters in some of the route types. This in turn stops
  "phantom constraints" from propagating all the way up the call
  stack, removing a few Send, Sync, and 'static requirements.
* I decoupled `State` from `RequestParams`, so route handlers now
  take two arguments. This makes the interfaces a bit more consistent
  between `at` (where the handler takes a `State` directly) and `get`,
  etc. (where the handler takes a `<State as ReadState>::State`.
* I ran into a Rust compiler bug which trips when we have Fn constraints
  that take a reference and return a future (which we do everywhere, now).
  There is a reasonable workaround that just requires a little bit more
  boxing. This should be thoroughly documented now.
* Improved documentation and expanded/added examples throughout.