Clean up handling of `Accept` header.
Currently, we construct an Accept header manually whenever we need
one using low-level functions, we require them to be mutable so we
can sort them on-demand, and we typically pass around `Option<Accept>`
even when `Accept` has an obvious default value (wildcard `*`).

This change adds a method to `RequestParams` to compute the `Accept`
header in a uniform way, automatically sorts it so it doesn't have
to be mutable after construction, and replaces `None` with a wildcard
so we don't have to deal with options.