Bundle the default public directory and serve it by default.
It is important to have a default public directory, because the
default HTML fragments for api.toml assume the existence of
`style.css` and some icons. This can lead to confusing behavior
where the user has not custom-configured anything, but the website
doesn't display correctly out of the box, unless this directory is
included whenever the default HTML is included.