Support prometheus endpoints
We add a new pseudo-method METRICS, similar to how we dispatch
SOCKET requests as if there were a SOCKET HTTP method. A request
is considered to have the METRICS method if its actual method is
GET and its preferred content type is text/plain. This is backwards
compatible since all existing endpoints return either application/json
or application/octet-stream.

A new middleware is installed on endpoints that have a METRICS
handler. It intercepts GET requests that are actually METRICS
requests and routes them to the corresponding METRICS handler.

Closes #70