Re-implement `StatusCode` to support bincode
`tide::StatusCode` uses `deserialize_any` in its deserialization,
which doesn't work with bincode. We implement a copy of their data
structure, but with a better serde implementation. Conversions are
provided between `StatusCode`, `tide::StatusCode`, and `u16`. In
addition, our enum has the exact same variant's as tide's and
supports the same public methods, so it should be backwards compatible.

A newtype wrapper would have required adding less code, but we would
have lost the fact that `StatusCode` is an enum, and matches would
have become far messier.