Support multiple versions of the same API
This gives us the capability of making backwards-compatible updates
to a system in production, which in turn reduces the pressure to get
all of our APIs perfect before a launch.

If we want to remove a problematic endpoint or make backwards-incompatible
changes after release, we can release a new major version of the API with
those changes, and continue to serve the old version under a version prefix.
We can tell existing clients to point at the old version of the API, so they
can continue to run for some period of time before updating their software.

This also allows us to do hotfixes in production. For example, suppose there
is some bug which we cannot fix without a breaking change to the API. We can
add a new route with a `/hotfix` prefix or suffix. Only the affected clients
need to change their code to use the hotfix endpoint. In the next major version,
the hotfix endpoint can replace the original.

Note that we can add new methods and make other backwards-compatible changes
with only a minor version bump, which can replace the existing API instead
of deploying a new version.

Closes #185