Put version prefix _before_ API prefix, instead of after
The rationale is that now a base URL like `https://hostname/v1` can
be passed into a client service, which can then access multiple API
modules from that URL, as long as all the modules support v1.
Different modules are still allowed to support different versions,
but this is not the common case.

This makes request parsing and routing a bit more complicated, but
things are greatly simplified by using a middleware to handle version
redirects. Now all the redirect code is isolated and separate from
the actual logic of handling routes.