Runtime binding for app-level format version
The App format version is no longer required to be consistent with
anything else (APIs all have their own versions), so we don't need it
to be a static parameter of the whole `App`. We can move it to just a
parameter of `serve`, which simplifies the interface a bit.

Also found, fixed, and added a test for a bug where we were using the
app version to serialize the response for API-level version endpoints,
rather than the API version.