Permit handling of socket messages to/from client in separate tasks
This allows a socket handler to spawn separate tasks to handle messages
to and from the client. This satisfies a use case of the node dashboard
server.

There are two ways to handle this, neither of which worked before this
change:
* split the connection into a separate stream and sink, send one to each
  task. This requires messages to be sent into the sink by value, due to
  a subtle lifetime issue that is explained in the documentation. Thus,
  this change adds an implementation of Sink<ToClient> for Connection, to
  complement the existing Sink<&ToClient> impl
* clone the connection and send one clone into each task. This change
  implements Clone for Connection. The Pin<Box<dyn Sink>> is stateless
  and derived from the WebSocketConnection. Thus, we can "clone" it by
  recreating it from a clone of the connection. All the state is cloned
  with the connection itself.

New documentation tests/examples are provided to demonstrate the two
methods of accomplishing this use case and prove they compile.